tech: Rename the logger with correct name
